core.utils refactoring v0.1
