+++ marketgraphicimages/core/new_password_reset_email.py
@@ -1,13 +1,14 @@
 from templated_mail.mail import BaseEmailMessage
 
-from .utils import (
+from .confirmation_code import (
     create_six_digit_confirmation_code,
     user_confirmation_code_to_db,
 )
 
 
 class PasswordResetEmail(BaseEmailMessage):
-    """Sends a confirmation code to the user."""
+    """Sends a confirmation code to the user.
+    Save confirmation code to db(UserConfirmationCode)."""
     template_name = "email/password_reset.html"
 
     def get_context_data(self) -> dict:
@@ -15,5 +16,4 @@ class PasswordResetEmail(BaseEmailMessage):
         user = context.get("user")
         context["token"] = create_six_digit_confirmation_code()
         user_confirmation_code_to_db(context["token"], user)
-        # print(context["token"])
         return context
